+++ image_classification_torch_loadgen_py/torch_loadgen_classifier.py
@@ -47,7 +47,7 @@ given_channel_stds          = [0.229, 0.224, 0.225]
 
 loader_object               = ImagenetLoader(preprocessed_imagenet_dir, MODEL_IMAGE_HEIGHT, MODEL_IMAGE_WIDTH, data_layout, normalize_symmetric, subtract_mean_bool, given_channel_means, given_channel_stds)
 preprocessed_image_buffer   = None
-preprocessed_image_map      = np.empty(dataset_size, dtype=np.int)   # this type should be able to hold indices in range 0:dataset_size
+preprocessed_image_map      = np.empty(dataset_size, dtype=np.int32)   # this type should be able to hold indices in range 0:dataset_size
 model                       = None
